Added Paragon button to CharacterFrame
Not sure if universally we want the button to be anchored to the top center of the player frame for everyone, regardless of their UI setups (ElvUI vs. non-ElvUI) since ElvUI displays the character frame differently.  We can either leave it as it with the button top center, or possibly check if the player has ElvUI enabled, then use the "TOPLEFT", 17, -17" anchor point and if they don't have ElvUI enabled, then use the "TOP", 0, 12" anchor point.