Carry over excess XP when leveling up
Previously when reaching the next Paragon level, it was always setting the XP back to 0.  If there was ever an excess/surplus of XP that was more than the max XP, the extra XP would be lost instead of rolling into the next level.

The Paragon button will no longer display on the Character frame if they have not yet unlocked the Paragon system (reached minPlayerLevel config value).

Also cleaned up the scripts a bit and tidied up some of the config checks.