Fix nil arithmetic value on first login
This nil arithmetic value error would only occur once on the first login because the database tables weren't populated yet.  After logging in for the first time, the database tables would be populated and the error would disappear, which is why I never saw the error.  But once I did a test on a fresh install I noticed the error.